Test setStatusMutation
