+++ uobtheatre/utils/validators.py
@@ -4,8 +4,6 @@ from typing import Generator, Union, Callable, Any, Optional
 from dataclasses import dataclass
 from uobtheatre.utils import exceptions
 
-import graphene
-
 
 @dataclass
 class ValidationError(exceptions.MutationException):
@@ -14,22 +12,12 @@ class ValidationError(exceptions.MutationException):
 
     def resolve(self):
         if self.attribute:
-            return exceptions.FieldError(
-                field=self.attribute, code=400, message=self.message
-            )
-        return exceptions.NonFieldError(code=400, message=self.message)
-
-
-"""
-        submit_draft
-        approve_pending -> approved (new)
-        publish_approved -> if they can edit
-"""
-
-
-class ValidationErrorNode(graphene.ObjectType):
-    message = graphene.String(required=True)
-    attribute = graphene.String()
+            return [
+                exceptions.FieldError(
+                    field=self.attribute, code=400, message=self.message
+                )
+            ]
+        return [exceptions.NonFieldError(code=400, message=self.message)]
 
 
 @dataclass
